requirements.txt
scikit-learn
pandas
flask
flask-cors
transformers
torch
setuptools
numpy==1.26.4
tf-keras
Pillow
ultralytics